added comment

+++ src/main/java/pl/akademiakodu/miniBlog/dao/PostDao.java
@@ -3,5 +3,6 @@ package pl.akademiakodu.miniBlog.dao;
 import org.springframework.data.repository.CrudRepository;
 import pl.akademiakodu.miniBlog.model.Post;
 
-public interface PostDao extends CrudRepository<Post, Long> {
+public interface PostDao extends CrudRepository<Post, Integer> {
+
 }
